requirements.txt
asgiref==3.7.2
bootstrap4==0.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.59.0
django-crispy-forms==2.1
django-encrypted-model-fields==0.6.5
django-formtools==2.5.1
django-otp==1.3.0
django-phonenumber-field==7.2.0
django-two-factor-auth==1.15.5
idna==3.6
oauthlib==3.2.2
phonenumberslite==8.13.27
pycparser==2.21
PyJWT==2.8.0
pypng==0.20220715.0
python-decouple==3.8
python3-openid==3.2.0
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
